Refactored out the connection.fs file.
This is the first stage of a clean up that should make the solution more scalable with a single bocket pool shared amongst all clients.
Still plenty of clean up needed, especially around the pattern matching which was a quick implementation.

Still not happy with the Start method of the BocketPool reiterating the collection adding in the callback, any better ideas anyone...